improved VirtualClock and command scheduler pipeline interaction
