+++ Domain/Scheduling/CommandScheduler.cs
@@ -148,83 +148,5 @@ namespace Microsoft.Its.Domain
             };
             return scheduledCommand;
         }
-
-        internal class CommandsInPipeline : IEnumerable<IScheduledCommand>
-        {
-            private readonly ConcurrentDictionary<IScheduledCommand, DateTimeOffset> commands = new ConcurrentDictionary<IScheduledCommand, DateTimeOffset>();
-         
-            public void Add(IScheduledCommand command)
-            {
-                var now = Clock.Now();
-                commands.AddOrUpdate(
-                    command,
-                    now,
-                    (c, t) => now);
-            }
-
-            public void Remove(IScheduledCommand command)
-            {
-                DateTimeOffset _;
-                commands.TryRemove(command, out _);
-            }
-
-            public async Task Done()
-            {
-                while (true)
-                {
-                    var now = Clock.Current;
-                    if (!commands.Keys.Any(c => c.IsDue(now)))
-                    {
-                        return;
-                    }
-                }
-            }
-
-            public IEnumerator<IScheduledCommand> GetEnumerator()
-            {
-                return commands.Keys.GetEnumerator();
-            }
-
-            IEnumerator IEnumerable.GetEnumerator()
-            {
-                return GetEnumerator();
-            }
-        }
-
-        public static Configuration TraceCommandsFor<TAggregate>(
-            this Configuration configuration)
-            where TAggregate : class, IEventSourced
-        {
-            var commandsInPipeline = TrackCommandsInPipeline(configuration);
-
-            return configuration.AddToCommandSchedulerPipeline<TAggregate>(
-                schedule: async (command, next) =>
-                {
-                    commandsInPipeline.Add(command);
-                    await next(command);
-                    Trace.WriteLine(Clock.Now() + " [Schedule] " + command);
-                },
-                deliver: async (command, next) =>
-                {
-                    await next(command);
-                    Trace.WriteLine(Clock.Now() + " [Deliver] " + command);
-                    commandsInPipeline.Remove(command);
-                });
-        }
-
-        internal static CommandsInPipeline TrackCommandsInPipeline(
-            Configuration configuration)
-        {
-            // resolve and register so there's only a single instance registered at any given time
-            CommandsInPipeline inPipeline;
-
-            lock (lockObj)
-            {
-                inPipeline = configuration.Container.Resolve<CommandsInPipeline>();
-                configuration.Container.Register(c => inPipeline);
-            }
-
-            return inPipeline;
-        }
     }
 }
